Fix lint: split Skit+Run, drop dead FileOutcome.destination, fix #require
- Remove .swift-version pin (unblocks swiftly toolchain resolution)
- Move CLI presentation helpers into Skit.Run+Render.swift to satisfy
  the 300-line file_length cap; add guarded `import Subprocess` back
- Drop assign-only `destination` from DirectoryRender.FileOutcome
- Use `try #require(...)` form in SkitSubprocessTimeoutTests
