--- .swift-version
@@ -1,1 +1,0 @@
-6.1
